Move stopped vehicle detector to its own helper file.
